# Show normal help before git describe for performance.