Feel free to clean this up. 

Basically, `git describe` can take a very long time (I've had it take up to 10 seconds), and it is annoying to wait for that long just to show the empty help command, which most programs display in under a second. 

Full description:

The time it takes to show `brew help` or just `brew` can take up to 10 seconds, compared to other programs which often take less than a second to show. 

The main issue with Brew's long startup time is `git describe`, which can take up to 9 seconds or so when not cached, and Ruby seems to take a second to start as well. See #2938.

Before change:
```
$ time brew help
Example usage:
  brew search [TEXT|/REGEX/]
  brew (info|home|options) [FORMULA...]
  brew install FORMULA...
  brew update
  brew upgrade [FORMULA...]
  brew uninstall FORMULA...
  brew list [FORMULA...]

Troubleshooting:
  brew config
  brew doctor
  brew install -vd FORMULA

Developers:
  brew create [URL [--no-fetch]]
  brew edit [FORMULA...]
  https://docs.brew.sh/Formula-Cookbook.html

Further help:
  man brew
  brew help [COMMAND]
  brew home

real	0m10.083s
user	0m0.818s
sys	0m0.686s
```

What I did is I made the shell script run the help command, reimplemented with `echo`, before `git describe` was called, if the user entered `brew`, `brew help`, `brew -?`, `brew --usage`, `brew --help`, or `brew -h`, without any other arguments, and then exit. 

After change:
```
$ time brew help
  Example usage:
    brew search [TEXT|/REGEX/]
    brew (info|home|options) [FORMULA...]
    brew install FORMULA...
    brew update
    brew upgrade [FORMULA...]
    brew uninstall FORMULA...
    brew list [FORMULA...]
  Troubleshooting:
    brew config
    brew doctor
    brew install -vd FORMULA
  Developers:
    brew create [URL [--no-fetch]]
    brew edit [FORMULA...]
    https://docs.brew.sh/Formula-Cookbook.html
  Further help:
    man brew
    brew help [COMMAND]
    brew home

real	0m0.143s
user	0m0.013s
sys	0m0.031s
```

As you can see, 0.143s is a lot less than 10.0183s.

This change will improve UX, as getting to the help screen faster is always an improvement, and this also matches most other programs in speed.

A potential improvement would be to run this for all `brew help` subcommands or simply not running `git describe` unless we actually use it.

Also, I made the function `showHelp`, but only called it once. That could be condensed, but it could be used for something else. I don't know. Again, feel free to put the echo statement directly into the if statement. 

- [x] Have you followed the guidelines in our [Contributing](https://github.com/Homebrew/brew/blob/master/CONTRIBUTING.md) document?
- [x] Have you checked to ensure there aren't other open [Pull Requests](https://github.com/Homebrew/brew/pulls) for the same change?
- [x] Have you added an explanation of what your changes do and why you'd like us to include them?
- [ ] Have you written new tests for your changes? [Here's an example](https://github.com/Homebrew/homebrew/pull/49031).
- [x] Have you successfully run `brew tests` with your changes locally?

-----