@MikeMcQuaid 2017-10-19
> The main issue with Brew's long startup time is git describe, which can take up to 9 seconds or so when not cached, and Ruby seems to take a second to start as well. See #2938.

A better alternative would be to speed up `git describe` rather than bypassing it in this way. I'm open to discussing more ways to do that but I'm not up for just bypassing it.