@reitermarkus 2017-10-19
> `git describe` can take a very long time

So it *can* take very long, but *when* would that be the case? I cannot reproduce the slowness of > 10 s, not even > 1 s.