@easyaspi314 2017-10-20
Well, here is some info:

My MacBook sucks, I don't think there is any denying that: 
>MacBook (13-inch, Mid 2009) - MacBook5,2
>Mac OS X 10.11.6 El Capitan
>2.13 GHz Intel Core 2 Duo
>4 GB 667 MHz DDR2 SDRAM
>NVIDIA GeForce 9400M 256 MB
>5400 RPM 1 TB hard drive

I use the latest git (2.14.2), but system git is pretty much the same speed.

I do have way too many formulae installed though, 212 to be exact, according to `brew list | wc -l`. 

```
$brew doctor
Please note that these warnings are just used to help the Homebrew maintainers
with debugging if you file an issue. If everything you use Homebrew for is
working fine: please don't worry and just ignore them. Thanks!

Warning: Putting non-prefixed coreutils in your path can cause gmp builds to fail. < note: I manually changed my path to be sure to remove coreutils from the PATH and it did nothing.

Warning: Unbrewed header files were found in /usr/local/include.
If you didn't put them there on purpose they could cause problems when
building Homebrew formulae, and may need to be deleted.

Unexpected header files:
  /usr/local/include/jack/control.h
  /usr/local/include/jack/intclient.h
  /usr/local/include/jack/jack.h
  /usr/local/include/jack/jslist.h
  /usr/local/include/jack/metadata.h
  /usr/local/include/jack/midiport.h
  /usr/local/include/jack/net.h
  /usr/local/include/jack/ringbuffer.h
  /usr/local/include/jack/session.h
  /usr/local/include/jack/statistics.h
  /usr/local/include/jack/systemdeps.h
  /usr/local/include/jack/thread.h
  /usr/local/include/jack/transport.h
  /usr/local/include/jack/types.h
  /usr/local/include/jack/uuid.h
  /usr/local/include/jack/weakjack.h
  /usr/local/include/jack/weakmacros.h

Warning: Unbrewed .pc files were found in /usr/local/lib/pkgconfig.
If you didn't put them there on purpose they could cause problems when
building Homebrew formulae, and may need to be deleted.

Unexpected .pc files:
  /usr/local/lib/pkgconfig/jack.pc

Warning: You have uncommitted modifications to Homebrew
If this is a surprise to you, then you should stash these modifications.
Stashing returns Homebrew to a pristine state but can be undone
should you later need to do so for some reason.
  cd /Users/easyaspi/Homebrew/Library && git stash && git clean -d -f

Warning: Your Homebrew's prefix is not /usr/local.
You can install Homebrew anywhere you want but some bottles (binary packages)
can only be used with a /usr/local prefix and some formulae (packages)
may not build correctly with a non-/usr/local prefix.
$ brew config                                                                                                                                                                                                                                          
HOMEBREW_VERSION: 1.3.6-dirty < with my commit on it
ORIGIN: https://github.com/Homebrew/brew.git
HEAD: 270b752f5d9d218bfbed6fe85b6974fa653fb25f
Last commit: 2 days ago
Core tap ORIGIN: https://github.com/Homebrew/homebrew-core
Core tap HEAD: 15e755fe33804f569f63bda2026fb515daa2b4f6
Core tap last commit: 12 hours ago
HOMEBREW_PREFIX: /Users/easyaspi/Homebrew
HOMEBREW_REPOSITORY: /Users/easyaspi/Homebrew
HOMEBREW_CELLAR: /Users/easyaspi/Homebrew/Cellar
HOMEBREW_BOTTLE_DOMAIN: https://homebrew.bintray.com
CPU: dual-core 64-bit penryn
Homebrew Ruby: 2.3.3 => /Users/easyaspi/Homebrew/Library/Homebrew/vendor/portable-ruby/2.3.3/bin/ruby
Clang: 8.0 build 800
Git: 2.14.2 => /Users/easyaspi/Homebrew/bin/git
Perl: /Users/easyaspi/Homebrew/bin/perl => /Users/easyaspi/Homebrew/Cellar/perl/5.26.0/bin/perl
Python: /Users/easyaspi/Homebrew/opt/python/libexec/bin/python => /Users/easyaspi/Homebrew/Cellar/python/2.7.14/Frameworks/Python.framework/Versions/2.7/bin/python2.7
Ruby: /Users/easyaspi/Homebrew/bin/ruby => /Users/easyaspi/homebrew/Cellar/ruby/2.4.2/bin/ruby
Java: 9, 1.8.0_25
macOS: 10.11.6-x86_64
Xcode: 8.2.1
CLT: 8.2.0.0.1.1480973914
X11: 2.7.9 => /opt/X11
$ brew list
ack	     bzip2	   diffutils	fontconfig    glib-networking		 gtkmm		     libao	   librsvg	  libyaml      ncdu	    pcre	      pyqt	     sdl			 valgrind	      zsh-lovers
adns	     cabextract    docbook	fontforge     glibmm			 gzip		     libassuan	   libsamplerate  little-cms2  ncurses	    pcre2	      python	     selenium-server-standalone  vim		      zsh-syntax-highlighting
ant	     cairo	   docbook-xsl	fortune       gmp			 harfbuzz	     libcroco	   libsigc++	  llvm	       ne	    perl	      python3	     sfml			 vimpager
apktool      cairomm	   dosfstools	freetype      gnu-getopt		 hicolor-icon-theme  libepoxy	   libsndfile	  llvm@3.9     net-snmp     pigz	      qt	     shared-mime-info		 webp
aspell	     ccache	   doxygen	gawk	      gnu-sed			 hr		     libffi	   libsoxr	  lua	       nettle	    pinentry	      qt5	     sip			 wget
astyle	     clang-format  dpkg		gd	      gnu-tar			 icu4c		     libfreenect   libssh2	  lv	       ninja	    pixman	      rbenv	     smartmontools		 wine
atk	     cmake	   emacs	gdbm	      gnupg			 imagemagick	     libgcrypt	   libtasn1	  lz4	       node	    pkg-config	      readline	     speex			 x264
atkmm	     colordiff	   exa		gdk-pixbuf    gnutls			 intltool	     libgpg-error  libtiff	  make	       npth	    portaudio	      repo	     speexdsp			 xmlto
autoconf     coreutils	   exiv2	gegl	      gobject-introspection	 itstool	     libgphoto2    libtool	  makedepend   nspr	    postgresql	      rgbds	     sphinx-doc			 xvid
automake     cowsay	   fdupes	gem-lolcat    gpatch			 jasper		     libicns	   libunistring   makeicns     openssl	    pulseaudio	      ruby	     sqlite			 xz
babl	     csfml	   ffind	gerrit-tools  graphite2			 jpeg		     libiconv	   libusb	  mono	       openssl@1.1  py2cairo	      ruby-build     swig			 yasm
bash	     curl	   ffmpeg	gettext       graphviz			 json-c		     libksba	   libusb-compat  moreutils    p11-kit	    pyenv	      rust	     texi2html			 zlib
berkeley-db  dash	   fish		gexiv2	      grep			 json-glib	     libmypaint    libvo-aacenc   most	       p7zip	    pyenv-virtualenv  s-lang	     timidity			 zsh
brew-gem     dbus	   flac		git	      gsettings-desktop-schemas  lame		     libogg	   libvorbis	  mpfr	       pango	    pygobject	      samba	     unrar			 zsh-autosuggestions
brotli	     dex	   fluid-synth	glib	      gtk+			 less		     libpng	   libxml2	  nano	       pangomm	    pygtk	      sane-backends  unzip			 zsh-completions

```