@MikeMcQuaid 2017-10-20
The number of formulae you have installed won't make any difference, just FYI.

> My MacBook sucks, I don't think there is any denying that:
> 5400 RPM 1 TB hard drive

Yep, that's the main culprit. A large, slow drive just plain sucks for macOS nowadays I'm afraid. Note if you set `HOMEBREW_UPDATE_TO_TAG` in your environment and run `brew update` to ensure you always stick on a tag you may find `git describe` ends up being much faster.